requirements.txt
langchain==0.3.7
langchain_community==0.3.7
langchain_core==0.3.20
langchain_ollama==0.2.0
matplotlib==3.9.2
neo4j==5.26.0
networkx==3.3
numpy==1.24.4
PyYAML==6.0.1
PyYAML==6.0.2
seaborn==0.13.2
streamlit==1.39.0
tqdm==4.66.4
